Fix race condition in unit tests
