+++ pkg/components/discover/watcher_proc_other.go
@@ -1,0 +1,12 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build !linux
+
+package discover
+
+// placeholder files to allow local compilation/unit testing in non-linux environments
+
+func currentTime() uint64 {
+	return 0
+}
